Clean up the selection-retention highlight
Three cleanups over the branch, no behaviour change:

- Reuse: the new module had copied the CSS Custom Highlight registry plumbing
  (feature detection, lazy registration, range swap) out of the page reader's
  word-highlight module almost verbatim. Extracted it to
  `lib/custom-highlight.ts` as `createSingleRangeHighlight(name)` and pointed
  both at it, dropping ~50 lines of duplication. `setWordHighlight` /
  `clearWordHighlight` keep their signatures, so the page reader and the
  extension that imports them are untouched.

- Dead state: `retainedRangeRef` was written in two places and never read once
  painting moved into `openToolbar`. Removed; the cloned range goes straight to
  the paint call.

- Theming: the highlight colour was a hardcoded blue. The neighbouring
  `::highlight(reader-word)` rule reads a custom property set from design tokens
  in `article-view.tsx`; follow that and set
  `--selection-retain-background-color` from `primaryColor.component3`, so a
  held passage matches the quote-share mark and follows the theme. The literal
  stays only as a fallback.
