+++ src/lib/selection-retention-highlight.ts
@@ -4,54 +4,25 @@
  * On touch devices the OS tears down the native text selection the instant the
  * user taps anything outside it — including our own selection toolbar. There is
  * no reliable way to stop that (preventing the default on the pointer-down does
- * not save the OS selection UI on mobile). So instead of fighting it, we keep a
- * snapshot of the chosen passage and re-paint it with a custom highlight while
- * the toolbar is in use. Like the page reader's word highlight, this paints over
- * a `Range` without mutating the DOM, so it never disturbs the article or the
- * live selection it stands in for. See the `::highlight(selection-retain)` rule
- * in `styles.css`.
+ * not save the OS selection UI on mobile). So instead of fighting it, we paint
+ * the chosen passage with a custom highlight for as long as the toolbar is up.
+ * `::selection` paints above custom highlights, so it stays invisible under the
+ * real selection and only takes over once the OS drops it.
+ *
+ * Painting over a `Range` never mutates the DOM, so this can't disturb the
+ * article or the live selection it stands in for. See the
+ * `::highlight(selection-retain)` rule in `styles.css`.
  */
 
-/** Registered name for the custom highlight (see the `::highlight()` rule). */
-const HIGHLIGHT_NAME = "selection-retain";
+import { createSingleRangeHighlight } from "#/lib/custom-highlight";
 
-interface HighlightLike {
-  add(range: Range): void;
-  clear(): void;
-}
-
-interface HighlightRegistryLike {
-  set(name: string, highlight: HighlightLike): void;
-  delete(name: string): void;
-}
-
-let activeHighlight: HighlightLike | null = null;
-
-function highlightRegistry(): HighlightRegistryLike | undefined {
-  return (globalThis.CSS as unknown as { highlights?: HighlightRegistryLike })
-    ?.highlights;
-}
+const highlight = createSingleRangeHighlight("selection-retain");
 
 /** Paint the retained selection; returns false when the browser lacks the API. */
 export function setSelectionRetentionHighlight(range: Range): boolean {
-  const registry = highlightRegistry();
-  const Ctor = (
-    globalThis as unknown as {
-      Highlight?: new (...ranges: Array<Range>) => HighlightLike;
-    }
-  ).Highlight;
-  if (!registry || !Ctor) return false;
-
-  if (!activeHighlight) {
-    activeHighlight = new Ctor();
-    registry.set(HIGHLIGHT_NAME, activeHighlight);
-  }
-  activeHighlight.clear();
-  activeHighlight.add(range);
-  return true;
+  return highlight.paint(range);
 }
 
 export function clearSelectionRetentionHighlight(): void {
-  highlightRegistry()?.delete(HIGHLIGHT_NAME);
-  activeHighlight = null;
+  highlight.clear();
 }
